Rename / Move of samples project
