+++ samples/NuGet.Server.V2.Samples.OwinHost/Controllers/NuGetPublicODataController.cs
@@ -7,7 +7,7 @@ using System.Text;
 using System.Threading.Tasks;
 using System.Net.Http;
 
-namespace NuGet.Server.V2.OwinSampleHost
+namespace NuGet.Server.V2.Samples.OwinHost.Controllers
 {
     /// <summary>
     /// Controller that exposes Program.NuGetPublicRepository as NuGet OData feed
