+++ samples/NuGet.Server.V2.Samples.OwinHost/Properties/AssemblyInfo.cs
